Books/TextMiningABC/requirements.txt
matplotlib
mecab-python-windows
numpy
pyQt5
cairocffi
nltk